Clean up the HFNC dashboard (#309)
